Adds workspace struct and simple resize
Workspace struct represents a virtual workspace such that they can
eventually be mapped in key/value pair and swapped between quite easily.
Each workspace embeds a primary window as well as a list of subsidiary
windows with the intention to eventually tile the subsidiaries in a
stack.

For now, when a window is added to the one (1) workspace, it's assigned
the primary position and full screened, and subsidiary windows are just
sort of existing in the nether region, but this allows you to test
sloppy focus.